% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_osm_city_data.R
\name{getOSMPlace}
\alias{getOSMPlace}
\title{Get the osm data for a city}
\usage{
getOSMPlace(
  country = NULL,
  placename = NULL,
  type = NULL,
  key = NULL,
  value = NULL
)
}
\arguments{
\item{placename}{The name of a place (Can be a country, a region or a city...)}

\item{type}{Either country, region or city}
}
\description{
Get the osm data for a city
}
